Improve exports for the dots_indicator
